SongDetailsFragment + MovieDetailsFragment + Parsers in separate classes + Aync task Callbacks
