+++ app/src/main/java/com/geoculturedemo/nickstamp/geoculturedemo/Callback/OnMovieDetailsDownloaded.java
@@ -1,8 +1,10 @@
 package com.geoculturedemo.nickstamp.geoculturedemo.Callback;
 
+import com.geoculturedemo.nickstamp.geoculturedemo.Model.Movie;
+
 /**
  * Created by nickstamp on 1/30/2016.
  */
 public interface OnMovieDetailsDownloaded {
-    public void onSongDetailsDownload();
+    void onDownload(Movie movie);
 }
