+++ app/src/main/java/com/geoculturedemo/nickstamp/geoculturedemo/Callback/OnSongDetailsDownloaded.java
@@ -1,8 +1,11 @@
 package com.geoculturedemo.nickstamp.geoculturedemo.Callback;
 
+import com.geoculturedemo.nickstamp.geoculturedemo.Model.Movie;
+import com.geoculturedemo.nickstamp.geoculturedemo.Model.Song;
+
 /**
  * Created by nickstamp on 1/30/2016.
  */
 public interface OnSongDetailsDownloaded {
-    public void onSongDetailsDownload();
+    public void onDownload(Song song);
 }
