+++ app/src/main/java/com/geoculturedemo/nickstamp/geoculturedemo/Fragment/SongFragment.java
@@ -23,9 +23,11 @@ import android.widget.TextView;
 
 import com.daimajia.androidanimations.library.Techniques;
 import com.daimajia.androidanimations.library.YoYo;
+import com.geoculturedemo.nickstamp.geoculturedemo.Callback.OnSongDetailsDownloaded;
 import com.geoculturedemo.nickstamp.geoculturedemo.Database.Database;
 import com.geoculturedemo.nickstamp.geoculturedemo.GeoCultureApp;
 import com.geoculturedemo.nickstamp.geoculturedemo.Model.Song;
+import com.geoculturedemo.nickstamp.geoculturedemo.Parser.SongDetailsParser;
 import com.geoculturedemo.nickstamp.geoculturedemo.R;
 import com.geoculturedemo.nickstamp.geoculturedemo.Utils.AnimationUtils;
 import com.geoculturedemo.nickstamp.geoculturedemo.Utils.FontUtils;
@@ -39,16 +41,18 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SongFragment extends Fragment implements View.OnClickListener {
+public class SongFragment extends Fragment implements View.OnClickListener, OnSongDetailsDownloaded {
 
     private static final String ARG_SONG = "song";
+    private static final String ARG_OFFLINE = "offline";
+
+    private boolean isOffline;
+
     private Song song;
 
     private TextView tvLyricsBy, tvTitle, tvMusicBy, tvLyrics;
     private LinearLayout llArtists;
     private ProgressBar pbArtists, pbLyrics;
-
-    List<String> artists, links;
 
     private Typeface typeface;
     private View fragmentView;
@@ -59,16 +63,15 @@ public class SongFragment extends Fragment implements View.OnClickListener {
     private SongDetailsParser songDetailsParser;
 
     public SongFragment() {
-        artists = new ArrayList<>();
-        links = new ArrayList<>();
 
     }
 
-    public static SongFragment newInstance(Song song1) {
+    public static SongFragment newInstance(Song song1, boolean isOffline) {
 
         SongFragment fragment = new SongFragment();
         Bundle args = new Bundle();
         args.putSerializable(ARG_SONG, song1);
+        args.putBoolean(ARG_OFFLINE, isOffline);
         fragment.setArguments(args);
         return fragment;
     }
@@ -78,6 +81,7 @@ public class SongFragment extends Fragment implements View.OnClickListener {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
             song = (Song) getArguments().getSerializable(ARG_SONG);
+            isOffline = getArguments().getBoolean(ARG_OFFLINE);
         }
     }
 
@@ -110,8 +114,10 @@ public class SongFragment extends Fragment implements View.OnClickListener {
 
             database = ((GeoCultureApp) getActivity().getApplication()).getDatabase();
 
-            songDetailsParser = new SongDetailsParser();
-            songDetailsParser.execute();
+            if (!isOffline) {
+                songDetailsParser = new SongDetailsParser(song, this);
+                songDetailsParser.execute();
+            }
         }
 
 
@@ -143,168 +149,90 @@ public class SongFragment extends Fragment implements View.OnClickListener {
         songDetailsParser.cancel(true);
     }
 
-    public class SongDetailsParser extends AsyncTask<Void, Void, Void> {
 
-        private String lyrics;
+    @Override
+    public void onDownload(final Song song) {
 
-        public SongDetailsParser() {
+        this.song.setArtist(song.getArtist());
+        this.song.setLinks(song.getLinks());
+        this.song.setLyrics(song.getLyrics());
 
+        AnimationUtils.crossfade(llArtists, pbArtists);
+        AnimationUtils.crossfade(tvLyrics, pbLyrics);
 
-        }
+        tvLyrics.setText(this.song.getLyrics());
 
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
+        String[] artists = this.song.getArtist().split("\\|");
+        String[] links = this.song.getLinks().split("\\|");
+        for (int i = 0; i < artists.length; i++) {
+            final String currentArtist = artists[i];
 
-        }
-
-        @Override
-        protected Void doInBackground(Void... params) {
-
-            //try to connect 3 times
-            int tries = 0;
-            boolean success = false;
-
-            while (tries < 3 && !success) {
-                tries++;
-
-                try {
-                    // Connect to the web site
-                    Document document = Jsoup.connect(song.getUrl()).get();
-
-                    Element table = document.select(".row3").get(0);
-
-                    Element singerElement = table.getElementsByClass("singers").get(0);
-                    Elements trSingers = singerElement.getElementsByTag("tr");
-                    for (Element e : trSingers) {
-                        String singer = e.text();
-                        artists.add(singer);
-                        for (int i = 0; i < e.getElementsByTag("td").size(); i++) {
-                            String s = e.getElementsByTag("td").get(i).getElementsByTag("a").attr("href");
-                            if (s.contains("youtube")) {
-                                links.add(s);
-                                break;
-                            }
-                        }
-                        if (artists.size() > links.size())
-                            links.add("");
-                    }
-
-                    StringBuilder sbArtists = new StringBuilder(), sbLinks = new StringBuilder();
-                    for (String s : artists) {
-                        sbArtists.append(s).append("|");
-                    }
-                    for (String s : links) {
-                        sbLinks.append(s).append(" ");
-                    }
-                    sbArtists.deleteCharAt(sbArtists.length() - 1);
-                    sbLinks.deleteCharAt(sbLinks.length() - 1);
-                    song.setArtist(sbArtists.toString());
-                    song.setLinks(sbLinks.toString());
-
-                    Elements divs = table.getElementsByTag("div");
-                    divs.remove();
-
-                    Elements b = table.getElementsByTag("b");
-                    if (b != null)
-                        b.remove();
-
-
-                    lyrics = Html.fromHtml(table.html()).toString();
-                    song.setLyrics(lyrics);
-
-                    success = true;
-
-                } catch (IOException e) {
-                    e.printStackTrace();
-                } catch (IndexOutOfBoundsException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            return null;
-
-        }
-
-        @Override
-        protected void onPostExecute(Void aVoid) {
-            super.onPostExecute(aVoid);
-
-            AnimationUtils.crossfade(llArtists, pbArtists);
-            AnimationUtils.crossfade(tvLyrics, pbLyrics);
-
-            tvLyrics.setText(lyrics);
-
-
-            for (int i = 0; i < artists.size(); i++) {
-                final String currentArtist = artists.get(i);
-
-                TextView tv = new TextView(getContext());
-                tv.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
-                tv.setText(currentArtist);
-                typeface = Typeface.createFromAsset(getContext().getAssets(), "fonts/Roboto-Regular.ttf");
-                tv.setTypeface(typeface);
+            TextView tv = new TextView(getContext());
+            tv.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
+            tv.setText(currentArtist);
+            typeface = Typeface.createFromAsset(getContext().getAssets(), "fonts/Roboto-Regular.ttf");
+            tv.setTypeface(typeface);
 
                 /*int padding_in_dp = 8;  // 12 dps
                 final float scale = getResources().getDisplayMetrics().density;
                 int padding_in_px = (int) (padding_in_dp * scale + 0.5f);
                 tv.setPadding(0, padding_in_px, 0, padding_in_px);*/
 
-                int[] attrs = new int[]{R.attr.selectableItemBackground};
-                TypedArray typedArray = getActivity().obtainStyledAttributes(attrs);
-                int backgroundResource = typedArray.getResourceId(0, 0);
-                tv.setClickable(true);
-                tv.setBackgroundResource(backgroundResource);
-                typedArray.recycle();
+            int[] attrs = new int[]{R.attr.selectableItemBackground};
+            TypedArray typedArray = getActivity().obtainStyledAttributes(attrs);
+            int backgroundResource = typedArray.getResourceId(0, 0);
+            tv.setClickable(true);
+            tv.setBackgroundResource(backgroundResource);
+            typedArray.recycle();
 
-                tv.setGravity(Gravity.CENTER_VERTICAL);
-                if (links.get(i).trim().length() > 0) {
-                    final String link = links.get(i);
+            tv.setGravity(Gravity.CENTER_VERTICAL);
+            if (links[i].trim().length() > 0
+                    && !links[i].equals("-")) {
+                final String link = links[i];
 
-                    tv.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(link)));
-                        }
-                    });
-                    tv.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.mipmap.ic_tube, 0);
-                    tv.setCompoundDrawablePadding(4);
-                } else {
+                Log.i("nikos", "link:" + link);
+
+                tv.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(link)));
+                    }
+                });
+                tv.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.mipmap.ic_tube, 0);
+                tv.setCompoundDrawablePadding(4);
+            } else {
 
 
-                    tv.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
+                tv.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
 
-                            Intent intent = new Intent(Intent.ACTION_SEARCH);
-                            intent.setPackage("com.google.android.youtube");
+                        Intent intent = new Intent(Intent.ACTION_SEARCH);
+                        intent.setPackage("com.google.android.youtube");
 
-                            String s = currentArtist.substring(2, currentArtist.length());
+                        String s = currentArtist.substring(2, currentArtist.length());
 
-                            intent.putExtra("query", song.getTitle() + " " + s);
-                            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                            startActivity(intent);
+                        intent.putExtra("query", song.getTitle() + " " + s);
+                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        startActivity(intent);
 
-                        }
-                    });
-                    tv.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.mipmap.ic_search, 0);
-                    tv.setCompoundDrawablePadding(4);
-                }
-
-                tv.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, LinearLayout.LayoutParams.WRAP_CONTENT));
-                llArtists.addView(tv);
+                    }
+                });
+                tv.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.mipmap.ic_search, 0);
+                tv.setCompoundDrawablePadding(4);
             }
 
-            isSaved = database.isSaved(song);
-            if (isSaved)
-                fab.setImageResource(R.drawable.ic_star);
-            else
-                fab.setImageResource(R.drawable.ic_star_outline);
-            fab.show();
+            tv.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+            llArtists.addView(tv);
         }
 
+        isSaved = database.isSaved(song);
+        if (isSaved)
+            fab.setImageResource(R.drawable.ic_star);
+        else
+            fab.setImageResource(R.drawable.ic_star_outline);
+        fab.show();
     }
-
 }
